Add ability to record time in created plates
Driver can now add a plate to the system and let the driver start the
meter. Next time the driver enters the plate numbers and such plates are
already in the system and the meter is running the driver will be informed
about that fact.
Also added new tests.
Added new entry to markdown file.