Add ablity to stop meter for plate
Added "stopAndPay" post method, which allows driver to stop the meter for
a given plate.
Now after entering the same plate app wil treat it as if it were a new
plate. What we care for in this functionality is if are there any plates
for which the meter is running. This allows the operator to search for
plates that are running or not.
Also added a few tests.