Add Pricing functionality + Tests
Added functionality that calculates the difference in time the meter
started and was ended and multiplies it by propper currency value. The
computed fee is presented to the driver based on the currency that the
driver chose. After 'paying', the fee is stored as a Plate parameter.
Added required elements:
Service - Pricing
Abstract class Pricing
its heir that extends the Pricing - PLNPricing
Enumeration that defines regions - Region
Required Mock Tests